日更

+++ TiHMS/ServerModule/src/main/java/tiquartet/ServerModule/po/UserPO.java
@@ -111,7 +111,7 @@ public class UserPO implements Serializable{
 		return this.birthday;
 	}
 	
-	public void setcredit(int credit){
+	public void setcredit(double credit){
 		this.credit=credit;
 	}
